# backend/docs/openapi.yaml
openapi: "3.1.0"
info:
  title: NyumbaLink API
  version: "2.0.0"
  description: |
    ## NyumbaLink — AI-Powered Accommodation Brokerage System
    **Dar es Salaam, Tanzania**

    NyumbaLink connects **Seekers** (tenants) with verified **Landlords** across
    Dar es Salaam's five municipalities — Kinondoni, Ilala, Ubungo, Temeke, and Kigamboni.

    ### Authentication
    Protected endpoints require a **Bearer JWT** in the `Authorization` header.
    ```
    Authorization: Bearer <access_token>
    ```
    Obtain tokens via `POST /api/v1/auth/login`.
    Access tokens expire after **24 hours**.
    Use `POST /api/v1/auth/refresh` to renew.

    ### Roles
    | Role | Description |
    |---|---|
    | `SEEKER` | Tenant searching for accommodation |
    | `LANDLORD` | Property owner managing listings |
    | `ADMIN` | Platform administrator |

    ### Base URL
    - Development: `http://localhost:8000`
    - Production: `https://api.nyumbalink.co.tz`
  contact:
    name: NyumbaLink Support
    email: mnemosadick@gmail.com
    url: https://nyumbalink.co.tz
  license:
    name: MIT License
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.nyumbalink.co.tz
    description: Production server
security:
  - BearerAuth: []
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT access token. Obtain via `POST /api/v1/auth/login`.
        Include as: `Authorization: Bearer <token>`
  schemas:
...
    RegisterRequest:
      type: object
      required: [first_name, last_name, email, password]
      properties:
        first_name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Sadick"
        last_name:
          type: string
          minLength: 1
          maxLength: 100
          example: "Mnemonics"
        email:
          type: string
          format: email
          example: "sadick@example.co.tz"
        phone:
          type: string
          nullable: true
          example: "+255653870753"
        password:
          type: string
          minLength: 8
          example: "SecurePass123!"
        role:
          type: string
          enum: [SEEKER, LANDLORD, ADMIN]
          default: SEEKER
          example: "SEEKER"
        language_preference:
          type: string
          enum: [en, sw]
          default: "en"
          example: "en"
    LoginRequest:
      type: object
      required: [email, password]
      properties:
        email:
          type: string
          format: email
          example: "sadick@example.co.tz"
        password:
          type: string
          example: "SecurePass123!"
        role:
          type: string
          enum: [SEEKER, LANDLORD, ADMIN]
          nullable: true
          example: "SEEKER"
    TokenPair:
      type: object
      properties:
        access:
          type: string
          description: JWT access token (expires 24h)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
        refresh:
          type: string
          description: JWT refresh token (expires 30 days)
          example: "eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9..."
    UserOut:
      type: object
      properties:
        id:
          type: string
          format: uuid
          example: "550e8400-e29b-41d4-a716-446655440000"
        first_name:
          type: string
          nullable: true
          example: "Sadick"
        last_name:
          type: string
          nullable: true
          example: "Mnemonics"
        email:
          type: string
          example: "sadick@example.co.tz"
        phone:
          type: string
          nullable: true
          example: "+255653870753"
        role:
          type: string
          enum: [SEEKER, LANDLORD, ADMIN]
          example: "SEEKER"
        profile_image:
          type: string
          nullable: true
          example: "https://cdn.nyumbalink.co.tz/avatars/user.jpg"
        language_preference:
          type: string
          example: "en"
        email_verified:
          type: boolean
          example: false
        phone_verified:
          type: boolean
          example: false
        two_factor_enabled:
          type: boolean
          example: false
        account_status:
          type: string
          enum: [ACTIVE, SUSPENDED, BANNED]
          example: "ACTIVE"
        created_at:
          type: string
          format: date-time
          example: "2025-01-15T10:30:00Z"
    AuthResponse:
      type: object
      properties:
        user:
          $ref: "#/components/schemas/UserOut"
        tokens:
          $ref: "#/components/schemas/TokenPair"
    UpdateProfileRequest:
      type: object
      properties:
        first_name:
          type: string
          nullable: true
        last_name:
          type: string
          nullable: true
        phone:
          type: string
          nullable: true
        language_preference:
          type: string
          enum: [en, sw]
          nullable: true
        profile_image:
          type: string
          nullable: true
    ChangePasswordRequest:
      type: object
      required: [current_password, new_password]
      properties:
        current_password:
          type: string
          example: "OldPass123!"
        new_password:
          type: string
          minLength: 8
          example: "NewPass456!"
    ForgotPasswordRequest:
      type: object
      required: [email]
      properties:
        email:
          type: string
          format: email
          example: "sadick@example.co.tz"
    ResetPasswordRequest:
      type: object
      required: [token, password]
      properties:
        token:
          type: string
          description: Signed JWT token from the reset email
        password:
          type: string
          minLength: 8
          example: "NewSecurePass789!"

    # ── Properties ────────────────────────────────────────────────────────────
...
    HTTPError:
      type: object
      properties:
        detail:
          type: string
          example: "Resource not found."
    ValidationError:
      type: object
      properties:
        detail:
          type: array
          items:
            type: object
            properties:
              loc:
                type: array
                items:
                  type: string
              msg:
                type: string
              type:
                type: string

# ── Tags ──────────────────────────────────────────────────────────────────────
...
paths:
...
  /api/v1/auth/register:
    post:
      tags: [Auth]
      summary: Register a new user
      description: |
        Create a new account. Supported roles: `SEEKER`, `LANDLORD`, `ADMIN`.
        Landlord accounts automatically receive a landlord profile record.
        Returns JWT access and refresh tokens immediately.
      security: []
      operationId: auth_register
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RegisterRequest"
            examples:
              seeker:
                summary: Register as Seeker
                value:
                  first_name: "Amina"
                  last_name: "Hassan"
                  email: "amina@example.co.tz"
                  phone: "+255712345678"
                  password: "SecurePass123!"
                  role: "SEEKER"
              landlord:
                summary: Register as Landlord
                value:
                  first_name: "John"
                  last_name: "Massawe"
                  email: "john@example.co.tz"
                  phone: "+255754001001"
                  password: "LandlordPass456!"
                  role: "LANDLORD"
      responses:
        "201":
          description: Account created successfully
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "409":
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ValidationError"
  /api/v1/auth/login:
    post:
      tags: [Auth]
      summary: Login and receive JWT tokens
      description: |
        Authenticate with email and password.
        Optionally specify `role` to enforce role matching.
        Returns access token (24h) and refresh token (30 days).
      security: []
      operationId: auth_login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/LoginRequest"
            examples:
              seeker_login:
                summary: Seeker login
                value:
                  email: "amina@example.co.tz"
                  password: "SecurePass123!"
                  role: "SEEKER"
              admin_login:
                summary: Admin login
                value:
                  email: "admin@nyumbalink.co.tz"
                  password: "AdminPass789!"
                  role: "ADMIN"
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthResponse"
        "401":
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
        "403":
          description: Account suspended or role mismatch
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/HTTPError"
  /api/v1/auth/me:
    get:
      tags: [Auth]
      summary: Get current user profile
      description: Returns the authenticated user's full profile.
      operationId: auth_get_me
      responses:
        "200":
          description: User profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
        "401":
          description: Not authenticated
    patch:
      tags: [Auth]
      summary: Update current user profile
      description: Update first name, last name, phone, language, or profile image.
      operationId: auth_update_me
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProfileRequest"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserOut"
  /api/v1/auth/change-password:
    post:
      tags: [Auth]
      summary: Change password
      description: Change password for the currently authenticated user.
      operationId: auth_change_password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ChangePasswordRequest"
      responses:
        "204":
          description: Password changed
        "400":
          description: Current password incorrect
  /api/v1/auth/forgot-password:
    post:
      tags: [Auth]
      summary: Request password reset email
      description: |
        Sends a signed password-reset JWT to the registered email address.
        Always returns 204 (even if email not found) to prevent enumeration.
        **Note:** Email sending must be wired in production.
      security: []
      operationId: auth_forgot_password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ForgotPasswordRequest"
      responses:
        "204":
          description: Reset email sent (or silently ignored if email not found)
  /api/v1/auth/reset-password:
    post:
      tags: [Auth]
      summary: Reset password using signed token
      description: Validates the JWT from the reset email and updates the password.
      security: []
      operationId: auth_reset_password
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ResetPasswordRequest"
      responses:
        "204":
          description: Password reset successfully
        "400":
          description: Invalid or expired token
  /api/v1/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Exchange a valid refresh token for a new access + refresh token pair.
      security: []
      operationId: auth_refresh
      parameters:
        - name: refresh
          in: query
          required: true
          schema:
            type: string
          description: The refresh token
      responses:
        "200":
          description: New token pair
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TokenPair"
        "401":
          description: Invalid or expired refresh token

  # ────────────────────────────── PROPERTIES ─────────────────────────────────
...